reworked country service and controller

+++ src/main/java/dev/cyan/travel/controller/CountryController.java
@@ -24,22 +24,39 @@ public class CountryController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Optional<Country>> getSingleCountry(@PathVariable ObjectId id) {
-        return new ResponseEntity<>(countryService.singleCountry(id), HttpStatus.OK);
+    public ResponseEntity<Country> getSingleCountry(@PathVariable ObjectId id) {
+        Optional<Country> country = countryService.singleCountry(id);
+        if (country.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>(country.get(), HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<Country> createCountry(@RequestBody Map<String, String> payload) {
-        return new ResponseEntity<>(countryService.createCountry(payload.get("name")), HttpStatus.CREATED);
+    public ResponseEntity<Country> createCountry(@RequestBody Country country) {
+        return new ResponseEntity<>(countryService.createCountry(country.getName()), HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Optional<Country>> updateCountry(@PathVariable ObjectId id, @RequestBody Country country) {
-        return new ResponseEntity<>(countryService.updateCountry(id, country), HttpStatus.OK);
+    public ResponseEntity<Country> updateCountry(@PathVariable ObjectId id, @RequestBody Country country) {
+        Optional<Country> existingCountry = countryService.singleCountry(id);
+        if (existingCountry.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        existingCountry.get().setName(country.getName());
+
+        return new ResponseEntity<>(countryService.saveCountry(existingCountry.get()), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Country> deleteCountry(@PathVariable ObjectId id) {
+        Optional<Country> country = countryService.singleCountry(id);
+        if (country.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         countryService.deleteCountry(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
